2023photoelectricity1/requirements.txt
absl-py==1.4.0
backcall==0.2.0
cachetools==4.2.4
certifi==2023.5.7
charset-normalizer==2.0.12
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
future==0.18.3
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.19.1
google-auth-oauthlib==0.4.6
grpcio==1.48.2
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
install==1.3.5
ipython==7.16.3
ipython-genutils==0.2.0
iso8601==1.1.0
jedi==0.17.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib @ file:///home/buwai/yolov5-seg/matplotlib-3.3.4-cp36-cp36m-linux_aarch64.whl
numpy==1.19.5
oauthlib==3.2.2
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.14.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
smmap==5.0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch @ file:///home/buwai/yolov5-seg/torch-1.10.0-cp36-cp36m-linux_aarch64.whl
torchvision @ file:///home/buwai/yolov5-seg/torchvision-0.11.0a0%2Bfa347eb-cp36-cp36m-linux_aarch64.whl
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.0.3
zipp==3.6.0
